agl: Small code refactor based on Windows port

+++ packages/agl/src/common/aglShaderProgramArchive.cpp
@@ -44,7 +44,7 @@ void createShaderProgram(
             it.getIndex(),
             variation.getName(),
             variation.getID(),
-            variation.ref().mValueNum,
+            static_cast<s32>(variation.ref().mValueNum),
             heap
         );
